Move around certifications tests
